adk/requirements.txt
litellm
google-adk
ipython
